just copy the source to the esm dist
right? why babel it if we don't need to? It was way bigger, and also why
the heck do you need to compile esm if the file is using esm and you're
targetting esm?

I'm using to dual publishing from TS source where you need a transpile
for both ESM and CJS